move StringHttpRequest etc into the restHttp project rather than salesforce
