+++ handlers/sf-contact-merge/src/test/scala/com/gu/sf_contact_merge/getsfcontacts/GetSfAddressEffectsTest.scala
@@ -3,7 +3,7 @@ package com.gu.sf_contact_merge.getsfcontacts
 import com.gu.effects.{GetFromS3, RawEffects}
 import com.gu.salesforce.SalesforceAuthenticate.SFAuthConfig
 import com.gu.salesforce.dev.SFEffectsData
-import com.gu.salesforce.{JsonHttp, SalesforceClient}
+import com.gu.salesforce.SalesforceClient
 import com.gu.sf_contact_merge.Types.IdentityId
 import com.gu.sf_contact_merge.getaccounts.GetZuoraContactDetails.EmailAddress
 import com.gu.sf_contact_merge.getsfcontacts.ToSfContactRequest.WireResult
@@ -11,6 +11,7 @@ import com.gu.sf_contact_merge.getsfcontacts.WireContactToSfContact.Types._
 import com.gu.util.resthttp.RestOp._
 import com.gu.test.EffectsTest
 import com.gu.util.config.{LoadConfigModule, Stage}
+import com.gu.util.resthttp.JsonHttp
 import org.scalatest.{FlatSpec, Matchers}
 import scalaz.\/-
