+++ lib/salesforce/src/main/scala/com/gu/salesforce/SalesforceClient.scala
@@ -1,6 +1,7 @@
 package com.gu.salesforce
 
 import com.gu.salesforce.SalesforceAuthenticate.{SFAuthConfig, SalesforceAuth}
+import com.gu.util.resthttp.JsonHttp.StringHttpRequest
 import com.gu.util.resthttp.RestRequestMaker._
 import com.gu.util.resthttp.{HttpOp, LazyClientFailableOp}
 import okhttp3.{Request, Response}
@@ -25,20 +26,5 @@ object SalesforceClient {
     val url = sfAuth.instance_url + requestInfo.relativePath.value
     authedBuilder.url(url).build()
   }
-
-  sealed trait RequestMethod {
-    def builder: Request.Builder
-  }
-  case class PostMethod(body: BodyAsString) extends RequestMethod {
-    override def builder: Request.Builder = new Request.Builder().post(createBodyFromString(body))
-  }
-  case class PatchMethod(body: BodyAsString) extends RequestMethod {
-    override def builder: Request.Builder = new Request.Builder().patch(createBodyFromString(body))
-  }
-  case object GetMethod extends RequestMethod {
-    override def builder: Request.Builder = new Request.Builder().get()
-  }
-
-  case class StringHttpRequest(requestMethod: RequestMethod, relativePath: RelativePath, urlParams: UrlParams)
 
 }
